MAINT: New benchmarks
