+++ asv.conf.json
@@ -65,7 +65,7 @@
 
     // The Pythons you'd like to test against.  If not provided, defaults
     // to the current version of Python used to run `asv`.
-    "pythons": ["3.11", "3.12"],
+    // "pythons": ["3.11", "3.12"],
 
     // The list of conda channel names to be searched for benchmark
     // dependency packages in the specified order
